A lot more spring cleaning stuff:     Adding comments to sort header code into meaningful chunks.     File enums moved into File class for organization.     Documentation for File interface     More proper documentation all around.

+++ include/ht_debug.h
@@ -75,7 +75,7 @@ namespace Hatchit {
         std::string DebugSprintF(const char* format, const Args& ... args);
 
         template<class ... Args> 
-        int DebugPrintF(const char* format, const Args& ... args);
+        size_t DebugPrintF(const char* format, const Args& ... args);
     }
 }
